move lang_is_equal() into dashlive.utils package
as the lang_is_equal() function can be used by other parts of
the codebase, in addition to the server request handler, move it
to a higher level utils directory.
